src/domain/categories/controllers/categories/categories.controller.ts
import {
  Body,
  Controller,
  HttpCode,
  HttpStatus,
  Post,
  UseGuards,
  Req,
} from '@nestjs/common';
import { JwtAuthGuard } from 'src/shared/guards';
import { CreateCategoryDto } from '../../dtos';
import { CategoriesService } from '../../services/categories/categories.service';

@Controller('categories')
export class CategoriesController {
  constructor(private readonly categoriesService: CategoriesService) {}

  @Post()
  @HttpCode(HttpStatus.CREATED)
  @UseGuards(JwtAuthGuard)
  async create(@Req() { userId }: any, @Body() category: CreateCategoryDto) {
    return await this.categoriesService.create({ ...category, userId });
  }
}
